use TextEncoder and TextDecoder when available
This commit allows the RxPlayer to use the `TextEncoder` and `TextDecoder`
APIs when available respectively to encode JS strings into an UTF-8
bytes sequence (TextEncoder doesn't seem to be able to encode into any
other encoding) and to decode from either UTF-8, UTF-16BE or UTF-16LE
into a JS string.

Because `TextEncoder` and `TextDecoder` are not defined in old browser
versions we claim to support and in IE11, we still fallback to custom
implementation either if it doesn't exist or if the operation fails.

It is important to note of a sensible difference between using
the `TextDecoder` interface and the previous implementation: when
encountering invalid byte sequences in the correponding encoding,
the `TextDecoder` will replace those by a "REPLACEMENT CHARACTER" (�).

This seems fine and even desirable, but the previous implementation just
threw in that same situation.
This means that we now have two different behaviors, depending on the
current platform / browser.

Those functions using the `TextDecoder` APIs are even directly defined
in the `StringUtils` tools, and thus that new behavior can be directly
noticable by applications using it.
Thankfully, nothing is defined in our API documentation about invalid
sequences.

Even if we can consider that this does not break our API (though it is
still unclear to me), it should be is something to keep in mind as this
might be unexpected for users relying on this API throwing.

Also, I tried to add unit tests, but it appears that "jsdom", on which
relies jest to perform unit test while simulation a browser in node,
does not include either APIs yet. Though it is under way:
https://github.com/jsdom/whatwg-encoding/pull/11