api: add minAutoBitrates feature to define a minimum adaptive bitrate
This commit is a proposal to add:
  1. a `minAudioBitrate` constructor option
  2. a `minVideoBitrate` constructor
  3. a `setMinAudioBitrate` method
  4. a `setMinVideoBitrate` method
  5. a `getMinAudioBitrate` method
  6. a `getMinVideoBitrate` method

Respectively mirroring the `maxAudioBitrate` and `maxVideoBitrate`
constructor option and the `setMaxAudioBitrate`, `setMaxVideoBitrate`,
`getMaxAudioBitrate` and `getMaxVideoBitrate` but for setting a lower
limit (instead of a higher limit) to qualities reachable through
adaptive streaming.

I added those because it makes perfect sense to have them in our APIs,
it fills a real need in my opinion and it does not deteriorate the
quality of the code much (the major downside being that new lines are
added to handle those options) as the logic to handle it is almost
exactly the same than for the maximum counterpart.