Avoid loading segment in a loop when in "native" textTrackMode
This commit fixes a regression brought in v3.22.0 where text track
segments would be loaded in a loop when in the native `textTrackMode`
(they would still be displayed at the right time, the issue just being a
lot of segment requests).

_Thankfully most (if not all) applications use the HTML textTrackMode.
As we do not test the native mode before release normally, this issue
was not spotted sooner._

The issue is due to a discrepancy between the format of the text track
data parsed by `transports` and the one accepted by the
`NativeTextSegmentBuffer`: the latter expected a `timescale` property
which was not communicated by the former (as the `timescale` is always
assumed to be equal to `1` since the last release).

This was not catched by TypeScript because there's some `any` type
between the two (to facilitate the creation of generic `PeriodStream`s,
each with its own generic `SegmentBuffer` linked to it).

I fixed it with a longer term solution than just fixing the
corresponding lines:
 now both text SegmentBuffers directly rely on the type defined by
`transports`. There's still the `any` between the two so there's still
no automatic check of the coherency, but at least for now we're sure
there's no difference in format between both places.