compat: throw a better error when no EME API is found
While testing playback of encrypted contents in an HTTP page on Edge
(on Edge and Chrome, EME APIs are only available on an HTTPS page), I
noticed that the error message was not explicit at all:

`"No isTypeSupported on MediaKeys"`.

This is because we were in a specific case, where the RxPlayer tried to
ponyfill the `requestMediaKeySystemAccess` API with another since-
deprecated API, `MediaKeys.isTypeSupported`.

This commit now checks if the `MediaKeys.isTypeSupported` API exists
 before fallbacking to that ponyfill, and no longer throw that error.
Now when neither the `requestMediaKeySystemAccess` nor the
`MediaKeys.isTypeSupported` API exist, the following error is thrown
instead:

`"This browser seems to be unable to play encrypted contents currently (are you using HTTPS?)"`