Fix Uint8Array believed to be an ArrayBuffer when blacklisting based on key-id
Since at least the v3.21.1 release, we had an issue that appeared when
blacklisting Representation based on the keyID (e.g. when a
`keystatuseschange` event indicates us that we are "output-restricted
and when fallbacking is enabled through the related `keySystems`
options).

This issue prevented from fallbacking, it just threw an Error message
along the lines of:
`TypeError: DataView: expected ArrayBuffer, got Uint8Array`

The root cause was linked to typing issues (we consequently cleverly
redirect all wrong-doing to typescript, not our fault at all here :p:
https://github.com/microsoft/TypeScript/issues/42534) where an
Uint8Array was considered to be a valid ArrayBuffer.

At some point, a `new
DataView(whatIThoughtToBeAnArrayBufferWhichIsInRealityAUint8Array)` was
done, leading to the aforementioned error.

The fix here is to now consider the type to be an Uint8Array and to do
Uint8Array things instead (i.e. no DataView anymore).