metaplaylist: set the right time and end for segments
When setting the `time` and `end` property on metaplaylist segment
objects, we have to apply an offset corresponding to the corresponding
content's start.

It was here mistakenly happening twice when calling `getSegments`:
  - once in the `_cloneWithPrivateInfos` private methods (which also
    apply to init segments, where we most likely don't want to apply
    offsets)
  - once directly in the `getSegments` method

This was a pretty big issue for metaplaylist contents but everything
still seemed to work fine most of the time, strangely enough.
This is surely because things like the segment's tfdt ISOBMFF box has
prevalence on the `time` property.

This issue was not present in our previous releases so this is not a bug
fix to add to the changelog.