improve content switching performance by loading the Manifest ASAP
The previous RxPlayer logic waited for a possible previous content to be
"cleaned-up" (its buffer removed, previous encryption sessions closed if
the right options are set etc.) before considering a new content.

We found that in some target, that clean-up can take time. Depending on
the options set and the type of contents (e.g. with multiple keys and
options to close all sessions) we noticed up to 500ms only to clean-up
the previous content.

This PR attempts to optimize that time by fetching the Manifest early
on, even before stopping the previous content.

Sadly this imply to move the `ManifestFetcher` from the `Init` to the
`API`, which now looks odd there.
Because there was no reason to now let the `SegmentFetcherCreator` in
the `Init` I moved it to the `API` as well.

This had the nice side-effect of globally removing options from the
`Init` (as many of those were only there to creates those two
instances).

But now, having some instances created in the `Init` (most of all the
`ABRManager`) and some of those in the `API` may be difficult to follow.
I don't really know what to do about that for now.

Also, there was no point to define the `fetchManifest` option directly
in the `InitializeOnMediaSource` function instead of the now only place
where it is used: the `manifestUpdateScheduler`. So I moved it there.

I tested that commit and it works well.

Thoughts?