doc: move architecture documentation files closer to the code
As discussed, I found that having the architecture documentation in
`doc/architecture` meant it was poorly accessible and might lead to it
being not updated properly.

This commit moves all architecture doc file closer to the code it
documents, as `README.md` files so they can even be easily seen by
outsiders who just casually browse our code in GitHub's interface.

It consequently removes the `doc/architecture` which means those pages
won't be generated in our HTML documentation.

However, this last point is not a big loss for now, the HTML
documentation being mainly to document the API.

Fixes #764