skip init segment's pssh boxes when encryption data is in the Manifest
EME APIs - to play encrypted contents - necessitate the communication of
"initialization data". This usually takes the form of a PSSH - metadata
found on initialization segment - but can more often than not also be
found in a Manifest file.

The previous RxPlayer behavior was to merge both the initialization
data found in the initialization segment and in the Manifest, before
communicating it.
This meant waiting for the initialization segment to be downloaded, so
we can merge them when both are known.

Yet in most cases, the same initialization data is available in both
places. Waiting for the initialization data to be loaded would here just
be an unnecessary delay added to decryption negociations, which is
usually the last step before a content starts to load.

This commit only uses the initialization data found in a Manifest when
there, and if not, it uses the initialization data in the initialization
segment.

It could potentially break cases where different initialization data is
defined in the Manifest and in the initialization segment, where the
initialization data actually wanted is in the latter.
But this should be an extremely rare case.