communicate separated initialization data to the EMEManager
This commit refactors the way encryption initialization data is
communicated to the `EMEManager` - the module calling DRM-related APIs.

Before that commit, the player emitted the concatenation of all
encryption data it found either in the Manifest or in initialization
segments once the latter has been loaded.
The `EMEManager` received this event, filtering it out if the exact same
data has already been found.

The data sent had the following format:
```ts
export interface IContentProtection { type : string; // initDataType
                                      data : Uint8Array; } //
concatenation of all encryption data
```

With this commit, the data sent is much more detailed.
It separates distinct encryption data (instead of just concatenating
them) and it includes which "system ID" are concerned - an ID allowing
to identify the key system in question - for each of these.

The data sent to the `EMEManager` now looks like:
```ts
/** Segment protection sent by the RxPlayer to the EMEManager. */
export interface IContentProtection {
  /**
   * Initialization data type.
   * String describing the format of the initialization data sent
     through this
   * event.
   * https://www.w3.org/TR/eme-initdata-registry/
   */
  type: string;
  /** Every initialization data for that type. */
  values: Array<{
    /**
     * Hex encoded system id, which identifies the key system.
     * https://dashif.org/identifiers/content_protection/
     */
    systemId: string;
    /**
     * The initialization data itself for that type and systemId.
     * For example, with "cenc" initialization data found in an ISOBMFF
       file, this
     * will be the whole PSSH box.
     */
     data: Uint8Array;
  }>;
}
```

Note that encryption initialization data is still concatenated before
being communicated through the `generateRequest` EME API, but it is now
done inside the `EMEManager` code.

---

The point of that commit is that it seems both cleaner (the `Manifest`
structure already stored this data like this anyway and concatenating
make no much sense there) and also allows future optimizations.

The main future optimization I'm thinking of is the ability to just rely
on the Manifest's data (and thus, not waiting for an initialization
segment to be loaded) when we know that it contains the initialization
data for the currently used key system (here, we would use the
`systemId` property to detect that).

---

In this commit, I encountered some issues with the `PersistentSessionsStore`
because it identified persisted MediaKeySession through the hash of the
complete initialization data.

Because not having retro-compatibility to persisted MediaKeySessions
might be a breaking change - even if we're talking about session
persisted in an older RxPlayer version - I chose to also perform the
concatenation there.