Single license per content implementation
 ## The need

 ### What is this?

This commit implements a `singleLicensePer` option, which allows an
application to tell the RxPlayer that the current content only has a
single license for the whole content, even when it has multiple keys
associated to different tracks or qualities.

The RxPlayer will then be able to perform a single license request and,
if the right options are set, to fallback from non-decryptable
contents, which will be both:

  1. keys which have the wrong key-status ("output-restricted" if
     `fallbackOn.keyOutputRestricted` is set and "internal-error" if
     `fallbackOn.keyInternalError` is set).

  2. keys which were not included in that license. Those will be
     fallbacked from, without needing to set any API.

The idea is also to be able to make evolutions to that API to implement
for example a single license per Period in the future.

 ### Why doing that?

The idea here is to perform a single license request, initially for a
single quality, which will in response return a license (technically
it could be multiple concatenated licenses in the CDM's point of
view depending of the key system, but it is still a single request and
data structure in the player's point of view) allowing to decrypt all
encrypted media from that same content.

The advantages of doing that are multiple. Most of all:

  1. it allows to only perform a single license request instead of
     multiple ones.
     Reducing possible server loads but also reducing network latencies.

  2. We could be able to tell much sooner which key are not supported
     and
     thus avoid switching to an undecipherable Representation to only
     later fallback from it.

  3. some other players not being compatible with contents necessitating
     multiple license requests, putting all keys in a single license can
     improve compatibility for when streams are used with multiple
     players.

  4. it reduces interactions with EME APIs, which we found to be very
     slow on some devices (we for example found out that generating a
     challenge can take more than 1 second on some embedded devices).

And surely many others.

Before that commit, the RxPlayer would already play those contents, but
would still perform multiple license requests reducing the potential
gains.

 ### What it means technically

Technically, the description of this feature is simple:

When `singleLicensePer` is set to `"content"`:

  - we should only perform a single license request, event when there's
    multiple initialization data encountered in the MPD and
initialization
    segments

  - we should fallback from `Representation`s whose corresponding
    key ids are either:
      1. Not signaled in the license
      2. Signaled in the license but have to be fallbacked from
         according to the `fallbackOn` options

 ## Implementation

 ### Re-purposing `handledInitData`

To implement that, the main idea is to re-purpose the `handledInitData`
cache (in the `EMEManager`), which stores initialization data that has
already been encountered.

The only goal of that structure before being to avoid creating multiple
MediaKeySessions for the same initialization data, it isn't a stretch to
redefine it as a structure avoiding creating multiple MediaKeySessions
linked to the same key ids.

I changed the way it is used in two ways:

  1. When `singleLicensePer` is set to `"content"` we can simply check
     when it's empty to only create a session on the first
     initialization data encountered (we can imagine also adding several
     IDs to that structure to easily implement `"period"` etc.).

  2. As we still need to fallback from any initialization data that has
     no corresponding key in the license when the latter is loaded, I
     had to add to that structure an Observable emitting `"blacklisted"`
     key ids from the corresponding `MediaKeySession` and the other
     key ids, now considered by opposition as `"whitelisted"`.

     Simply said, every key ids linked to a MediaKeySession which are
not
     present blacklisted (because of the usual `fallbackOn` options)
will
     be whitelisted.

     By listening to that Observable everytime new initialization data
is
     encountered, we can just compare the linked key id with whitelisted
     key ids that Observable emits - which will be once the license is
     pushed.
     If the license has already been pushed, the Observable will emit
     immediately.

 ### Adding the key id to `contentProtections$` event

Because we are using key ids here to detect which keys are not in a
given license, we have to emit those alongside the initialization data
to
the `EMEManager`.
This is done through the already-existing `contentProtections$`
Observable.

In the hypothesis that no key id is anounced in the Manifest, we may
thus have to parse initialization data to extract it from there, so
we can support multi-key per license mode with such contents.
This is not done here for the moment, but I'm working on it now.

In the meantime, initialization data for which the corresponding key id
is not known will still lead to another license request, which I felt
was
a good compromise.

 ## Does it work?

I tested it with success on:

  - Widevine L1 on STBs
  - Widevine L3 on Linux - Chrome
  - Widevine L3 on Linux - Firefox
  - Widevine L3 on Mac - Chrome
  - Widevine L3 on Windows - Edge
  - PlayReady SL3000 on Windows - Edge

I know it doesn't work yet on some set-top boxes, but this seems more an
exception than the norm. We've still scheduled a meeting to learn more
about this issue.
Depending on the result, we may want to perform updates on that code.