Mark decipherable Representation as decipherable even if they were un-decipherable before
This commit is based on the multiple-keys-per-license commits (in #904)
and add on top of it a new behavior.

Representations which have their key-id added to one of the
`MediaKeyStatusMap` linked to the current content and which are not
fallbacked from, will now have their `decipherable` property updated to
`true` - even if it was set to `false` before (and thus even if we had
fallbacked from it in the past).

This development was made a lot easier by the work done to support the
`singleLicensePer` option (#904) with now the concept of "whitelisted
key ids", in opposition of the "blacklisted key ids" we fallback from.

When a key id is found to be whitelisted (technically this means
currently that its `MediaKeyStatus` is not either: `"internal-error"`,
`"output-restricted"` or `"expired"`, the last one depending on
`keySystems options) we will now:

  - update the related `Representation`'s `decipherable` property to
    `true` - regardless of its previous state

  - schedule a `decipherabilityUpdate` event with that update through
    the API (and through the `Manifest` object with the event of the
    same name)

  - In the `AdaptationStream` - only if the updates changed the list
    of available Representations for that Adaptation, re-construct the
    list of the Representations the ABR has to choose from

---

To note that this has never been tested in real conditions.