Fixes on handling sudden fall in bandwidth
The RxPlayer has several mechanisms to adapt the current chosen video
and audio quality to the bitrate.

Most of them rely on bandwidth calculations based on the time each
previous segment requests made, but we noticed that this is sometimes
not enough.

Sometimes the fall of bandwidth is so sudden and acute that we cannot
afford waiting for the current segment requests to finish (prerequisite
to the regular behavior).
For those cases, the RxPlayer has a second logic specifically when the
buffer is "starving" - which means that is has no or few data left in
its buffer.

When in "starvation mode", the RxPlayer will pay attention to the
progress made by current requests, and ask to cancel those to switch
in quality if that progression seems to slow.

---

The "starvation mode" feature is quite old, but I noticed two issues
that made it not function properly:

  1. we relied on the `bufferGap` to know when to enter in starvation
     mode. However, the `bufferGap` has a peculiarity (due to legacy
     reasons which may be linked to the API) of being set to
     `Infinity` when no buffer is available at all at the current
     position.

    This completely falsify the results here.

  2. To infer that a request takes too much time, the RxPlayer needs
     to find which requests, from all pending requests, concern the
     an immediately coming segment.
     Here we checked that that segment's start must be AT LEAST in
     300 milliseconds.
     What we most likely wanted to do is to check for a segment
     which is AT MOST in 300 milliseconds.

For that second point I even raised that limit to be 1.2 seconds, as
300 milliseconds looked really imminent.