Add the clock's wantedTimeOffest to the BufferGarbageCollector's
The `BufferGarbageCollector` is an observable which regularly clean
the media buffers based on the distance between the current position
and the media data.
Consequently, it needs to be given the current playing position to work
properly.

To that effect the `position` property of each clock tick was given to
it. This is good, but we forgot to add the `wantedTimeOffest` to it if
it exists.
The `wantedTimeOffest` property was added as a workaround for cases
where the media element's position (what is actually reflected by the
`position` property) is different from the wanted position.
This only happens for now when first loading the content, as there the
seek to the initial position cannot be performed until the
`loadedmetadata` event has been received.

The true impact of this error is difficult to know and may cause no
actual problem today.
It depends on when the `loadedmetadata` is received, which is in most
scenarios just after the initialization segments have been pushed,
which means that no data has been added yet. If no data has been added,
there's no data to remove, so no problem.

An issue with that has only been seen on Samsung TVs, when we already
had issues with the initial seek not being performed.
Here the issue was that loaded segments would be removed immediately
after being pushed in cases where they are far from the `0` position.