Only use concerned encryption systemId when found in the Manifest
This PR is based on #911 and alows a performance optimization where the
license request for encrypted contents can be performed sooner.

The idea is to:
  - determine, from the key system currently used, which encryption
    initialization data is needed based on this data's "systemId"
  - if the Manifest contains that data directly embedded within it, send
    only that data as soon as the content in question is considered for
download (instead of waiting for the initialization segment to be loaded
and parsed as we did before) and do not send other encryption data for
that same content
  - if the Manifest does not contain that data - or if we couldn't
    determine the system data - just send all encryption data like
before (from both the Manifest and the initialization segment, implying
that both requests have to be done before)

When encryption data is present in the Manifest, as it is most often the
case, this allows theoretically to perform the license request parallely
to the request of the initialization segment.
Before, we had to perform the latter before the former.

---

I tried a similar idea in #901 but the behavior proposed (just taking
only encryption data from the Manifest when it exists there) was judged
to be too much of a change when compared to older RxPlayer versions.

---

Important note: Sadly this optimization can break persistent sessions
persisted from an older RxPlayer version.

This is because the ID used to identify those, was the hash of the
concatenation of all initialization data.
As the whole point of that PR is to not wait for all initialization data
to be available, this PR completely break this logic.

I'm unsure if that should be considered as a breaking change, but it
could break some usage, for example when persistent sessions were used
to play encrypted contents offline (and thus were we cannot afford just
re-doing the request).

In doubt, this PR adds a
`keySystems[].licenseStorage.disableRetroCompatibility` boolean.
If persistent MediaKeySessions are used, we will only perform that
optimization (by emitting or not the `systemId`) if that boolean is set
to `true`.
I find this solution to not be perfect because this is very implicit
(the fact that we're hiding the `systemId` is set to `true` and not just
explicitely telling not to perform any optimization) but I considered it
still good enough!

Maybe another behavior to break for a v4?