DRM: add maxSessionCacheSize option to limit maximum number of stored EME sessions
Add the `maxSessionCacheSize` keySystems option, which will allows any
application to set its own value when the default value of `50` is too
high.

Previously, those relied on the `closeSessionsOnStop` property, which
only kept open MediaKeySessions linked to the current content (and kept
alive all of those).

This new property seems much more adapted to those usages where the
limitation is known.

Note however that this limit may be different depending on how many keys
can be found in the licence as this option does not take in account this
value for simplicity reasons.