Avoid seeking over discontinuities when the browser seeks back
This commit fixes a problematic behavior on Tizen (mostly Samsung TVs)
where we could have an infinite seeking loop.

This issue is due to the fact that after seeking to a specific position
Tizen may actually seek to the closest intra video frame as a performance
tweak.

This is an unexpected behavior (I would argue that it goes against the
HTML5 specification) which may actually lead to a huge problem that is
difficult to work-around.

---

Let's say that the RxPlayer seek at at a position `T1`.
After loading the first video media segment, samsung sees that the first
intra frame is actually at a position `T2`, which is inferior to `T1`.

```
                            segment
video |--------------------|====|======|
                          T2   T1

current position => T1
```

In that condition, the browser seeks back to T2.

Now let's imagine we loaded the audio segment, that starts at `T3`, a
position after `T2`.

```
                            segment
video |--------------------|===========|
                          T2

audio |----------------------|=========|
                            T3

current position => T2
```

Here, the RxPlayer will seek automatically to T3 because there's no
audio data at T2, the current position.

However, the browser will then bring us back to T2, because that's where
the closest intra video frame is.
And, you might have guessed it, the RxPlayer will again seek to T3.

We will this way keep alternating between T2 and T3 multiple times (we
only are able to go outside this infernal loop when we're lucky and the
next clock tick - an RxPlayer interval - starts after we've reached T3).

---

The main problem is that those seek are very difficult to detect (they
seem to happen after a delay, for example).

The solution I've found is to store the position we're currently seeking
to and then detecting if we're before that position when the seek is
finished, and to avoid skipping discontinuities in that case.

This seems to work.

Because this is not a risky behavior (usually we won't be before a
seeked position after we seeked), I chose to generalize this new
behavior, so any new platform with the same weird concepts will be
automatically handled.