generalize segment paddings to all devices
Paddings allows to avoid overwriting segments that are currently being
decoded as we encountered many decoding issues when doing so.

The downside of that is that we will in consequence take more time to
raise the video/audio quality when the bandwidth suddenly raise-up.

Based on that, we previously only enabled paddings on devices we knew
had problems with overwriting decoded segments.

However, we start seeing more and more devices with this problem.
Because of that, it may now be difficult to continue that same strategy.

An alternative would be to provide an option to enable/disable paddings
or even to be able to configure their sizes, but this would come at the
cost of complexifying the API.

The other alternative would be to just generalize those paddings.
If it is sufficiently low (1.2 seconds in this commit), its
inconvenients may not really be important and it has the added advantage
of avoiding to load segment that are very soon finished.