eme: only persist MediaKeySessions once its keys are known
Before this commit, persistent MediaKeySessions were added to the
`PersistentSessionsStore` (for later retrieval) as soon as possible:
when their `sessionId` property was known (meaning as soon as the
`generateRequest` answered).

This worked pretty well, but we found a persistence-related issue that
could profit from taking another strategy:

After loading a persistent MediaKeySessions (that has previously been
added to the `PersistentSessionsStore`), the RxPlayer immediately checks
the status of its keys through the `keyStatuses` property.

However, on some platforms, the `keyStatuses` property is not directly
populated once the `load` call is done, but only after a
difficult-to-predict delay (sometimes immediately after, sometimes more
than 10 milliseconds after).

This imply that we might now have to wait after loading a persistent
session just to be sure that all keys information have been added to it.

We limited that by only waiting a delay if the `keyStatuses` property is
empty, yet this pre-condition was frequent: any time the user loaded a
new content before the license request succeeded (but after the
`generateRequest` call) we would be legitimately in a case where the
`keyStatuses` property is actually empty.

To avoid penalizing too much legitimate cases, this commit adds a
persistent session to the `PersistentSessionsStore` only once at least
one key is added to its `keyStatuses` property.

This ensures that having an empty `keyStatuses` property will now be
unusual enough.
In that case, incurring a delay (let's say up to 100ms) will be much
less penalizing.