drm: In a singleLicensePer content mode, blacklist the first kid if not encountered
We had a bug in our implementation of the `singleLicensePer` API (not
yet released, hopefully), where we would not fallback from a quality
linked to a key id not found in the licence, only if that key id was the
first encountered.

This is because the check we (I) added to do that only did it on
subsequent key ids, not the first one that triggered the
license-fetching process.

I tried to add a fix but it is not as simple as it looks, because if
we're doing that in a `singleLicensePer`: "content"` mode, why not
always (as it would make sense to always fallback when the wanted key id
is not found in the corresponding license).

However I prefer not to do that due to both not break the API and the
possible compatibilities issues as explained in the code.

So I was left doing it only if `singleLicensePer` was not set to the
default value, and putting that code in `eme_manager.ts` instead of
where I would have preferred to put it, initially, in
`check_key_statuses.ts`.