manifest: Rename Period.getPlayableAdaptations to getSupportedAdaptations
This method was only called by the `TrackChoiceManager` and a recent fix
ultimately led to the `NO_PLAYABLE_REPRESENTATIONS` not being thrown
anymore, as the `TrackChoiceManager` would just ignore tracks for which
all Representations are marked as not decipherable.

This could make sense, but I prefer to postpone such behavior change for
a future release.

This commit re-set the old behavior by just checking the codec support
through the `isSupported` property.