Do not consider that there is an init segment when only a BaseURL is available
This commit fixes an issue I encoutered while investigating #945 where I
noticed that the player loaded two times in a row the same text segment
anounced in the tested MPD.

---

The reason is that the RxPlayer has a peculiar behavior when no
`SegmentBase`, `SegmentList` nor `SegmentTemplate` element are found to
be linked to a given Representation.

In that situation, it creates a `TemplateRepresentationIndex` with both
an initialization segment and a media segment linked to that
Representation's URL.

For example, let's consider the following AdaptationSet:
```
<AdaptationSet id="2" contentType="text" mimeType="text/vtt"
lang="tlh">
  <Role schemeIdUri="urn:mpeg:dash:role:2011" value="caption"/>
  <Representation id="nuqneh" bandwidth="0">
    <BaseURL>https://retservilo.eo/qapla.vtt</BaseURL>
  </Representation>
</AdaptationSet>
```

Based on the `BaseURL` element, we infer the Representation's URL to be
`https://retservilo.eo/qapla.vtt`.

And because there's no `SegmentList`, `SegmentBase` nor
`SegmentTemplate` here (nor in its parent elements), we fall on the
default behavior which is to create a `TemplateRepresentationIndex`,
and that there is both an initialization segment and a media segment
with that URL: "https://retservilo.eo/qapla.vtt".

---

This means that in the case where this `Representation` is chosen, we will
download 2 times that segment. Once as an initialization segment, the
second time as a media segment.

From there, the question is: why loading both? Shouldn't we only consider
a media segment but no initialization segment in that situation?

I think (it was before I worked on this project) the original reason
this was done was to be safe, because that segment CAN in the end contain
initialization data. The solution would be to loading it first as purely an
initialization segment (and parsing initialization information from it)
and then as purely a media segment, ensuring that we're not missing on
anything

But a better solution to me would be to consider it as a media
segment, and handle cases where the media segment also contain
initialization data.

---

For now, the only important initialization data I found that we should
parse there (there's less important ones, for eample the timescale
anounced in a mdhd ISOBMFF box, but I didn't bother yet) was a possible
PSSH box (to properly handle it then in the `RepresentationStream`).

Doing that was very straightforward, I didn't encounter any issue
there.

In the end, parsing PSSH from media segments might even help us to
better implement advanced encryption use cases such as key rotation
and key hierarchy as it seems to rely on this.